+++ tests/test_investigations.py
@@ -1,6 +1,7 @@
-from circleguard import ReplayPath, Mod
+import numpy as np
 
-from tests.utils import CGTestCase, DELTA, RES, THRESHOLD_STEAL
+from circleguard import ReplayPath, Mod, Detect, StealResultSim, StealResultCorr
+from tests.utils import CGTestCase, DELTA, RES
 
 class TestCorrection(CGTestCase):
     @classmethod
@@ -36,6 +37,14 @@ class TestSteal(CGTestCase):
         cls.stolen2 = ReplayPath(RES / "stolen_replay2.osr")
         cls.legit1 = ReplayPath(RES / "legit_replay1.osr")
         cls.legit2 = ReplayPath(RES / "legit_replay2.osr")
+        cls.time_shifted1 = ReplayPath(RES / "stolen" / "stolen-time-shifted-1.osr")
+        cls.time_shifted2 = ReplayPath(RES / "stolen" / "stolen-time-shifted-2.osr")
+
+    @staticmethod
+    def add_noise_and_positional_translation_to_replay(replay, pixel_offset, std_deviation):
+        mean = [pixel_offset, pixel_offset]
+        covariance_matrix = [[std_deviation, std_deviation],[std_deviation,std_deviation]]
+        replay.xy = replay.xy + np.random.default_rng().multivariate_normal(mean, covariance_matrix, len(replay.xy))
 
     def test_cheated(self):
         # taken from http://redd.it/bvfv8j, remodded replay by same user (CielXDLP) from HDHR to FLHDHR
@@ -43,7 +52,7 @@ class TestSteal(CGTestCase):
         r = list(self.cg.steal_check(replays))
         self.assertEqual(len(r), 1, f"{len(r)} results returned instead of 1")
         r = r[0]
-        self.assertTrue(r.similarity < THRESHOLD_STEAL, "Cheated replays were not detected as cheated")
+        self.assertLess(r.similarity, Detect.SIM_LIMIT, "Cheated replays were not detected as cheated")
 
         r1 = r.replay1
         r2 = r.replay2
@@ -59,12 +68,27 @@ class TestSteal(CGTestCase):
         self.assertEqual(later.replay_id, 2805164683, "Later replay id was not correct")
         self.assertEqual(r1.username, r2.username, "Replay usernames did not match")
 
-    def test_legitimate_replaypath(self):
+    def test_cheated_time_shift(self):
+        replays = [self.time_shifted1, self.time_shifted2]
+        r = list(self.cg.steal_check(replays, method=Detect.STEAL_SIM))[0]
+        self.assertAlmostEqual(r.similarity, 17.30112, delta=DELTA, msg="Similarity is not correct")
+
+        # STEAL_SIM is currently *not* able to detect time shifts. If this
+        # changes we want to know! :P
+        self.assertGreater(r.similarity, Detect.SIM_LIMIT)
+
+        # STEAL_CORR should be able to, though.
+        r = list(self.cg.steal_check(replays, method=Detect.STEAL_CORR))[0]
+        self.assertGreater(r.correlation, Detect.CORR_LIMIT, "Cheated replays were not detected as cheated")
+        self.assertAlmostEqual(r.correlation, 0.99764, delta=DELTA, msg="Correlation is not correct")
+
+
+    def test_legitimate(self):
         replays = [self.legit1, self.legit2]
         r = list(self.cg.steal_check(replays))
         self.assertEqual(len(r), 1, f"{len(r)} results returned instead of 1")
         r = r[0]
-        self.assertFalse(r.similarity < THRESHOLD_STEAL, "Legitimate replays were detected as stolen")
+        self.assertGreater(r.similarity, Detect.SIM_LIMIT, "Legitimate replays were detected as stolen")
 
         r1 = r.replay1
         r2 = r.replay2
@@ -89,7 +113,7 @@ class TestSteal(CGTestCase):
             results_num = num * (num - 1) / 2 # n choose k formula with k=2
             self.assertEqual(len(r), results_num, f"{len(r)} results returned instead of {results_num}")
             r = r[0]
-            self.assertTrue(r.similarity < THRESHOLD_STEAL, f"Cheated replays were not detected as cheated at num {num}")
+            self.assertLess(r.similarity, Detect.SIM_LIMIT, f"Cheated replays were not detected as cheated at num {num}")
 
             r1 = r.replay1
             r2 = r.replay2
@@ -104,3 +128,18 @@ class TestSteal(CGTestCase):
             self.assertEqual(earlier.replay_id, 2801164636, f"Earlier replay id was not correct at num {num}")
             self.assertEqual(later.replay_id, 2805164683, f"Later replay id was not correct at num {num}")
             self.assertEqual(r1.username, r2.username, f"Replay usernames did not match at num {num}")
+
+    def test_robustness_to_translation(self):
+        # copy replay to avoid any missahaps when we mutate the data
+        stolen2 = ReplayPath(self.stolen2.path)
+        replays = [self.stolen1, stolen2]
+        self.cg.load(stolen2)
+        TestSteal.add_noise_and_positional_translation_to_replay(stolen2, 10, 3)
+        results = list(self.cg.steal_check(replays, method=Detect.STEAL_CORR | Detect.STEAL_SIM))
+
+        self.assertEqual(len(results), 2, f"{len(results)} results returned instead of 2")
+        for r in results:
+            if isinstance(r, StealResultSim):
+                self.assertLess(r.similarity, Detect.SIM_LIMIT, "Cheated replays were not detected as cheated with sim")
+            if isinstance(r, StealResultCorr):
+                self.assertGreater(r.correlation, Detect.CORR_LIMIT, "Cheated replays were not detected as cheated with corr")